pyproject.toml
[project]
name = "amee"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=6.0.2",
    "django-cors-headers>=4.9.0",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "drf-spectacular>=0.29.0",
    "pillow>=12.1.1",
    "pytest>=9.0.2",
    "pytest-django>=4.11.1",
]
